+++ Source/ActionRogueLike/Private/STargetDummy.cpp
@@ -12,7 +12,7 @@ ASTargetDummy::ASTargetDummy()
 	AttributeComp->OnHealthChanged.AddDynamic(this, &ASTargetDummy::OnHealthChanged);
 }
 
-void ASTargetDummy::OnHealthChanged(AActor* InstigatorActor, USAttributeComponent* OwningComp, float NewHealth, float Delta)
+void ASTargetDummy::OnHealthChanged(AActor* InstigatorActor, USAttributeComponent* OwningComp, int32 NewHealth, int32 Delta)
 {
 
 	if(Delta < 0.0f)
