+++ Source/ActionRogueLike/Public/AI/SAICharacter.h
@@ -36,7 +36,7 @@ protected:
 	virtual void PostInitializeComponents() override;
 
 	UFUNCTION()
-	void OnHealthChanged(AActor* InstigatorActor, USAttributeComponent* OwningComp, float NewHealth, float Delta);
+	void OnHealthChanged(AActor* InstigatorActor, USAttributeComponent* OwningComp, int32 NewHealth, int32 Delta);
 
 	//note: UPawnSensingComponent is deprecated - we should be using the AI Perception system instead
