+++ Source/ActionRogueLike/Public/STargetDummy.h
@@ -34,7 +34,7 @@ protected:
 
 
 	UFUNCTION()
-	void OnHealthChanged(AActor* InstigatorActor, USAttributeComponent* OwningComp, float NewHealth, float Delta);
+	void OnHealthChanged(AActor* InstigatorActor, USAttributeComponent* OwningComp, int32 NewHealth, int32 Delta);
 
 
 };
